fix: add react query provider for storybook
